backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.128.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.13.0",
    "asyncpg>=0.29.0",
    "psycopg2-binary>=2.9.0",
    "pytest>=7.0.0",
    "httpx>=0.24.0",
    "pytest-asyncio>=0.21.0",
    "greenlet>=3.0.0",
    "aiosqlite>=0.19.0",
    "ollama>=0.1.0",
    "pytesseract>=0.3.13",
    "pdf2image>=1.17.0",
    "pillow>=10.0.0",
    "rapidfuzz>=3.0.0",
    "sentence-transformers>=2.0.0",
    "python-multipart>=0.0.9",
    "aiofiles>=23.0.0",
]
